+++ src/olmo_core/internal/experiment.py
@@ -322,7 +322,12 @@ def build_config(
 def launch(config: ExperimentConfig):
     log.info(config)
     assert config.launch is not None
-    config.launch.launch(follow=True, launch_timeout=5 * 60)
+    config.launch.launch(
+        follow=True,
+        launch_timeout=5 * 60,
+        #  step_timeout=30 * 60,  # hard timeout kills the job
+        step_soft_timeout=10 * 60,  # soft timeout only sends slack warning
+    )
 
 
 def launch_prep(config: ExperimentConfig):