+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added option to skip ranges of steps in the trainer.
+- Send a Slack notification when a Beaker job appears to be stuck.
+
+## [v2.3.0](https://github.com/allenai/OLMo-core/releases/tag/v2.3.0) - 2025-10-17
+
 ### Fixed
 
 - Fixed parsing username+password git remote URLs in `launch.beaker` module.